+++ src/main/java/site/bidderown/server/bounded_context/bid/controller/BidController.java
@@ -21,18 +21,6 @@ public class BidController {
 
     private final BidService bidService;
 
-    @PostMapping("/bid")
-    public String registerBid(@RequestBody BidRequest bidRequest, @AuthenticationPrincipal User user){
-        bidService.handleBid(bidRequest, user.getUsername());
-        return "redirect:/bid/list?itemId=" + bidRequest.getItemId();
-    }
-
-    @PostMapping("/api/v1/bid")
-    @ResponseBody
-    public Long createBid(@RequestBody BidRequest bidRequest, @AuthenticationPrincipal User user){
-        return bidService.handleBid(bidRequest, user.getUsername());
-    }
-
     @GetMapping("/bid/list")
     @PreAuthorize("isAuthenticated()")
     public String bidList(@RequestParam Long itemId, Model model, @AuthenticationPrincipal User user){
@@ -40,19 +28,5 @@ public class BidController {
         return "usr/bid/list";
     }
 
-    @GetMapping("/api/v1/bid/list")
-    @ResponseBody
-    public BidResponses bidListApi(@RequestParam Long itemId){
-        BidDetails bidDetails = bidService.getBidStatistics(itemId);
-        List<BidResponse> bids = bidService.getBids(itemId);
-        return BidResponses.of(bidDetails, bids);
-    }
 
-    @GetMapping("/api/v1/bid-ids")
-    @ResponseBody
-    public List<Long> getBidItemIds(
-            @AuthenticationPrincipal User user
-    ) {
-        return bidService.getBidItemIds(user.getUsername());
-    }
 }
