+++ src/main/java/site/bidderown/server/bounded_context/notification/controller/NotificationApiController.java
@@ -5,11 +5,10 @@ import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.User;
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.bind.annotation.RestController;
 import site.bidderown.server.bounded_context.notification.controller.dto.NotificationResponse;
 import site.bidderown.server.bounded_context.notification.entity.Notification;
 import site.bidderown.server.bounded_context.notification.service.NotificationService;
@@ -17,23 +16,28 @@ import site.bidderown.server.bounded_context.notification.service.NotificationSe
 import java.util.List;
 import java.util.stream.Collectors;
 
-@Controller
+@RestController
 @RequiredArgsConstructor
-public class NotificationController {
-
+public class NotificationApiController {
     private final NotificationService notificationService;
-
-    @GetMapping("/notifications")
+    @GetMapping("/api/v1/notifications")
     @PreAuthorize("isAuthenticated()")
-    public String notificationList(Model model) {
+    public List<NotificationResponse> notificationList() {
         String username = SecurityContextHolder.getContext().getAuthentication().getName();
         List<Notification> list = notificationService.getNotifications(username);
-        model.addAttribute("notifications",
-                list.stream().map(NotificationResponse::of)
-                        .collect(Collectors.toList())
-        );
 
-        return "usr/notification/list";
+        return list.stream().map(NotificationResponse::of).collect(Collectors.toList());
         // return dto 변환 코드 작성해야함
+    }
+
+    @GetMapping("/api/v1/notification-check")
+    public Boolean checkNotification(@AuthenticationPrincipal User user) {
+        if (user == null) return false;
+        return notificationService.checkNotRead(user.getUsername());
+    }
+
+    @PutMapping("/api/v1/notification/readAll")
+    public void readAllNotification() {
+        notificationService.readAll();
     }
 }
