+++ src/main/resources/templates/usr/bid/list.html
@@ -168,10 +168,6 @@
                 }).catch(error => console.log(error));
 
                 axios.post('/api/v1/bid', {
-                    auth: {
-                        username: this.username,
-                        password: ''
-                    },
                     itemId: this.itemId,
                     itemPrice: this.bidPriceInput
                 }).then(response => {
